move Curl_wait_ms() to curlx
